Add files via upload
These are the basic circuit diagrams of the encoder for the OOK transmitter and the decoder for the OOK receiver. The diagrams show how the PIC16F628 is connected as an encoder and decoder. Inputs 1-10 of the encoder are connected to the Abel infrared sensors. Input 11 is connected to the electret microphone sensor. On the decoder output "n" will receive the data supplied to input "n" of the encoder. The pull down resistors on the encoder ensure that logic level 0 is transmitted when an input is not connected to a sensor. The resistors on the decoder outputs are to protect the decoder from unexpectedly high voltages on the serial ports of the PC. Note that pin 4 of each PIC is held high (+5V) to inhibit the reset function. Pin 5 is connected to ground and pin 14 is connected to +5V. There are no connections to pins 3, 15 and 16.